requirements.txt
ultralytics
easyocr
streamlit
opencv-python-headless
numpy
pillow
torch